+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://peps.python.org/pep-0440/).
 
+## v0.1.0a9 (2026-02-25)
+
+### Feat
+
+- **evaluator**: add execution_id, team_id, round_number to EvaluationRequest and BaseMetric.evaluate
+
+### Fix
+
+- **evaluator**: fix InvalidMetricNoEvaluate test to return None instead of valid MetricScore
+
 ## v0.1.0a8 (2026-01-28)
 
 ### Feat